back/requirements/base.txt
Django==3.2.13
pandas==1.5.0
django-cors-headers==3.9.0
djangorestframework==3.12.4
gunicorn==20.1.0
requests==2.26.0
whitenoise==6.0.0
unipath==1.1
django-material-admin==1.8.5
django-simple-history
psycopg2-binary==2.8.6
pandas==1.5.0
geopy==2.2.0
django-filter==21.1
djangorestframework-simplejwt==4.8.0